add witness info
