+++ core/blockcache/block_cache.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"sync"
 
-	"encoding/json"
 	"strconv"
 
 	"github.com/iost-official/Go-IOS-Protocol/common"
@@ -13,7 +12,6 @@ import (
 	"github.com/iost-official/Go-IOS-Protocol/core/global"
 	"github.com/iost-official/Go-IOS-Protocol/db"
 	"github.com/iost-official/Go-IOS-Protocol/ilog"
-	"github.com/iost-official/Go-IOS-Protocol/vm/database"
 )
 
 type CacheStatus int
@@ -29,85 +27,6 @@ const (
 	Single
 	Virtual
 )
-
-type WitnessList struct {
-	activeWitnessList    []string
-	pendingWitnessList   []string
-	pendingWitnessNumber int64
-}
-
-// SetPending set pending witness list
-func (wl *WitnessList) SetPending(pl []string) {
-	wl.pendingWitnessList = pl
-}
-
-// SetPendingNum set block number of pending witness
-func (wl *WitnessList) SetPendingNum(n int64) {
-	wl.pendingWitnessNumber = n
-}
-
-// SetActive set active witness list
-func (wl *WitnessList) SetActive(al []string) {
-	wl.activeWitnessList = al
-}
-
-// Pending get pending witness list
-func (wl *WitnessList) Pending() []string {
-	return wl.pendingWitnessList
-}
-
-// Active get active witness list
-func (wl *WitnessList) Active() []string {
-	return wl.activeWitnessList
-}
-
-// SetPendingNum get block number of pending witness
-func (wl *WitnessList) PendingNum() int64 {
-	return wl.pendingWitnessNumber
-}
-
-// UpdatePending update pending witness list
-func (wl *WitnessList) UpdatePending(mv db.MVCCDB) error {
-
-	vi := database.NewVisitor(0, mv)
-
-	spn := database.MustUnmarshal(vi.Get("iost.vote-" + "pendingBlockNumber"))
-	if spn == nil {
-		return errors.New("failed to get pending number")
-	}
-	pn, err := strconv.ParseInt(spn.(string), 10, 64)
-	if err != nil {
-		return err
-	}
-	wl.SetPendingNum(pn)
-
-	jwl := database.MustUnmarshal(vi.Get("iost.vote-" + "pendingProducerList"))
-	if jwl == nil {
-		return errors.New("failed to get pending list")
-	}
-
-	str := make([]string, 0)
-	err = json.Unmarshal([]byte(jwl.(string)), &str)
-	if err != nil {
-		return err
-	}
-	wl.SetPending(str)
-
-	return nil
-}
-
-func (wl *WitnessList) LibWitnessHandle() {
-	wl.SetActive(wl.Pending())
-}
-
-func (wl *WitnessList) CopyWitness(n *BlockCacheNode) {
-	if n == nil {
-		return
-	}
-	wl.SetActive(n.Active())
-	wl.SetPending(n.Pending())
-	wl.SetPendingNum(n.PendingNum())
-}
 
 type BlockCacheNode struct {
 	Block        *block.Block
@@ -242,15 +161,10 @@ func NewBlockCache(baseVariable global.BaseVariable) (*BlockCacheImpl, error) {
 		bc.hmset(bc.linkedRoot.Block.HeadHash(), bc.linkedRoot)
 		bc.leaf[bc.linkedRoot] = bc.linkedRoot.Number
 
-		if !bc.stateDB.Checkout(string(lib.HeadHash())) {
-			return nil, errors.New("failed to state db")
-		}
-		if err := bc.linkedRoot.UpdatePending(bc.stateDB); err != nil {
+		if err := bc.updatePending(bc.linkedRoot); err != nil {
 			return nil, err
 		}
-
 		bc.linkedRoot.LibWitnessHandle()
-
 		ilog.Info("Witness Block Num:", bc.LinkedRoot().Number)
 		for _, v := range bc.linkedRoot.Active() {
 			ilog.Info("ActiveWitness:", v)
@@ -281,15 +195,6 @@ func (bc *BlockCacheImpl) Link(bcn *BlockCacheNode) {
 	}
 }
 
-func (bc *BlockCacheImpl) initHead(h *BlockCacheNode) error {
-
-	bc.head = h
-	if err := bc.updatePending(bc.head); err != nil {
-		return err
-	}
-	return nil
-}
-
 func (bc *BlockCacheImpl) setHead(h *BlockCacheNode) error {
 	h.CopyWitness(h.Parent)
 	if h.Number%common.VoteInterval == 0 {
@@ -308,6 +213,12 @@ func (bc *BlockCacheImpl) updatePending(h *BlockCacheNode) error {
 			ilog.Error("failed to update pending, err:", err)
 			return err
 		}
+		if err := h.UpdateInfo(bc.stateDB); err != nil {
+			ilog.Error("failed to update witness info, err:", err)
+			return err
+		}
+	} else {
+		return errors.New("failed to state db")
 	}
 	return nil
 }
@@ -350,11 +261,9 @@ func (bc *BlockCacheImpl) AddGenesis(blk *block.Block) {
 	bc.linkedRoot = NewBCN(nil, blk)
 	bc.linkedRoot.Type = Linked
 
-	if bc.stateDB.Checkout(string(blk.HeadHash())) {
-		bc.linkedRoot.UpdatePending(bc.stateDB)
+	if err := bc.updatePending(bc.linkedRoot); err == nil {
 		bc.linkedRoot.LibWitnessHandle()
 	}
-
 	bc.head = bc.linkedRoot
 	bc.hmset(bc.linkedRoot.Block.HeadHash(), bc.linkedRoot)
 	bc.leaf[bc.linkedRoot] = bc.linkedRoot.Number
